+++ lib/Exceptions/FileNotFoundException.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace Catalytic\SDK\Exceptions;
+
+use Exception;
+
+/**
+ * An exception to be thrown when a File is not found
+ */
+class FileNotFoundException extends NotFoundException
+{
+    public function __construct($message, Exception $previous = null)
+    {
+        parent::__construct($message, $previous);
+    }
+}